workers/steam/presence/requirements.txt
steamio==1.1.3
coloredlogs==15.0.1
python-dotenv==1.0.1
